Implemented ability to play 3x3 sliding tiles as well as a button for Settings.
Note that 4x4 sliding tiles is hard coded as default (lines 38 and 61 in StartingActivity), this can be changed manually to play 3x3, but we must somehow link settings with this functionality in order to have the player choose in game what version they want to play.

Also note the code supports 5x5 but we must create tile resources and uncomment the switch statement in order for it to work.